frontend/requirements.txt
streamlit
pathlib
pandas
plotly-express
requests